requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.22
celery==3.1.19
Django==1.8.7
django-celery==3.1.17
django-cors-headers==1.1.0
django-pagination-bootstrap==1.0.7
django-vanilla-views==1.0.4
djangorestframework==3.3.2
gunicorn==19.4.1
kombu==3.0.33
MySQL-python==1.2.5
pysphere==0.1.7
pytz==2015.7
redis==2.10.5
requests==2.9.1
